feat: create clearer methods in positionator for determining position
